refactor(workspace): Rename tools to commands and peel unified_diff
The old tools module was process/session management, not agent tools.
Rename it to commands so the workspace layer matches ARCHITECTURE, and
move unified-diff text helpers out of patch.rs.
